enable decoupling of persist & policies
the change is backwards compatible with any implemntors of the
`Aggregate` trait, or implementors of `AggregateManager` who don't
override the default impl of `AggregateManager::persist`.

This enables external implementors of `AggregateManager` to override how
policies are run - e.g. to make them eventually constistent, or to allow
policies to fail while allowing `persist`/`handle_command` to succeed.